Move the `ChatMembersReader` to the upper package.
