+++ server/src/main/java/io/spine/examples/chatspn/server/message/MessageEditingProcess.java
@@ -38,6 +38,7 @@ import io.spine.examples.chatspn.message.event.MessageEdited;
 import io.spine.examples.chatspn.message.event.MessageEditingFailed;
 import io.spine.examples.chatspn.message.rejection.EditingRejections.MessageContentCannotBeUpdated;
 import io.spine.examples.chatspn.message.rejection.MessageCannotBeEdited;
+import io.spine.examples.chatspn.server.ChatMembersReader;
 import io.spine.examples.chatspn.server.ProjectionReader;
 import io.spine.server.command.Command;
 import io.spine.server.event.React;
